+++ Runtime/NavMeshLink.cs
@@ -64,6 +64,10 @@ namespace Unity.AI.Navigation
         [SerializeField]
         int m_Area;
 
+#if UNITY_EDITOR
+        int m_LastArea;
+#endif
+        
         /// <summary> Gets or sets the type of agent that can use the link. </summary>
         public int agentTypeID
         {
@@ -283,7 +287,13 @@ namespace Unity.AI.Navigation
         }
 
         // ensures serialized version is up-to-date at run-time, in case it was not updated in the Editor
-        void Awake() => UpgradeSerializedVersion();
+        void Awake()
+        {
+            UpgradeSerializedVersion();
+#if UNITY_EDITOR
+            m_LastArea = m_Area;
+#endif
+        }
 
         void OnEnable()
         {
@@ -392,7 +402,9 @@ namespace Unity.AI.Navigation
 
             m_LastPosition = transform.position;
             m_LastRotation = transform.rotation;
-
+#if UNITY_EDITOR
+            m_LastArea = m_Area;
+#endif
             RecordEndpointTransforms();
 
             GetWorldPositions(out m_LastStartWorldPosition, out m_LastEndWorldPosition);
@@ -453,7 +465,7 @@ namespace Unity.AI.Navigation
 
             m_Width = Mathf.Max(0.0f, m_Width);
 
-            if (!NavMesh.IsLinkValid(m_LinkInstance))
+            if (!NavMesh.IsLinkValid(m_LinkInstance) && (m_LastArea != 1 || m_Area == 1))
                 return;
 
             UpdateLink();
@@ -466,8 +478,9 @@ namespace Unity.AI.Navigation
             {
                 AddTracking(this);
             }
+            
+            m_LastArea = m_Area;
         }
-
         void Reset()
         {
             UpgradeSerializedVersion();